Colour picker outline done
