+++ milestone3/src/components/ColorPicker/ColorPicker.css
@@ -1,0 +1,49 @@
+.block {
+  width: 75px;
+  height: 200px;
+}
+
+.complement {
+  display: flex;
+  align-items: flex-start;
+  width: 200px;
+  height: 40px;
+  margin-left: 30px;
+  margin-top: 20px;
+}
+
+.complement-color {
+  width: 100px;
+  height: 40px;
+  margin-left: 30px;
+}
+
+.swap {
+  background: gray;
+  width: 40px;
+  height: 40px;
+}
+
+.colorPicker {
+  display: flex;
+}
+
+.display {
+  display: block;
+  align-items: flex-start;
+  width: 400px;
+  height: 200px;
+  background: lightgrey;
+}
+
+.wheel {
+  position: absolute;
+  right: 100px;
+}
+
+.slider {
+  margin-left: 50px;
+  margin-top: 5px;
+  display: block !important;
+  width: 230px !important;
+}
